refactor: add coverage on lib
